requirements.txt
Flask
Flask-Cors
Pillow
numpy
tensorflow
python-dotenv
